split out rabbitmqadmin from rabbitmq
Use it's own module so that we can check if the rabbitmqadmin is installed.
